fix(pgqueue): widen retry counters to bigint to prevent overflow
retry_count/max_retries were 32-bit INT with only a CHECK (>= 0). A
pathological crash-loop could overflow the counter; once retry_count
wraps negative the DLQ exhaustion test becomes unreliable and a message
could dodge the DLQ.

Widen retry_count (channel msg, pub/sub sub, DLQ) and max_retries
(channel msg) to BIGINT, raising the ceiling from ~2.1e9 to ~9.2e18.
Fresh queues are born BIGINT; a forward-only v4 migration widens
pre-existing per-queue tables, skipping columns already BIGINT.

Move internal max_retries scan targets NullInt32 -> NullInt64; no public
API change (RetryCount/MaxRetries were already int).

Closes #135

+++ pgqueue.go
@@ -1014,7 +1014,7 @@ func (pq *Queue) createPubSubTables(
 			visibility_timeout TIMESTAMPTZ,
 			available_at TIMESTAMPTZ NOT NULL DEFAULT now(),
 			claim_id UUID,
-			retry_count INT NOT NULL DEFAULT 0 CHECK (retry_count >= 0),
+			retry_count BIGINT NOT NULL DEFAULT 0 CHECK (retry_count >= 0),
 			error_message TEXT,
 			UNIQUE(message_id, subscriber_id),
 			FOREIGN KEY (message_id)
@@ -1104,8 +1104,8 @@ func (pq *Queue) createChannelTables(
 			payload BYTEA NOT NULL,
 			created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
 			status TEXT NOT NULL DEFAULT '%s',
-			retry_count INT NOT NULL DEFAULT 0 CHECK (retry_count >= 0),
-			max_retries INT CHECK (max_retries IS NULL OR max_retries >= 0),
+			retry_count BIGINT NOT NULL DEFAULT 0 CHECK (retry_count >= 0),
+			max_retries BIGINT CHECK (max_retries IS NULL OR max_retries >= 0),
 			visibility_timeout TIMESTAMPTZ,
 			available_at TIMESTAMPTZ NOT NULL DEFAULT now(),
 			claim_id UUID,
@@ -1196,7 +1196,7 @@ func (pq *Queue) createDLQTable(
 			subscriber_id TEXT,
 			payload BYTEA NOT NULL,
 			failure_reason TEXT NOT NULL,
-			retry_count INT NOT NULL CHECK (retry_count >= 0),
+			retry_count BIGINT NOT NULL CHECK (retry_count >= 0),
 			moved_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
 			metadata JSONB
 		)`, pq.dlqTable(tableName))
